User

+++ isd-backend/src/main/java/com/example/isdbackend/model/Provider.java
@@ -10,7 +10,7 @@ import java.util.Set;
 public class Provider {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String name;
     private Integer deliveryPrice;
@@ -34,11 +34,11 @@ public class Provider {
         this.isActive = isActive;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
